requirements.txt
crewai>=0.28.0
langchain-google-genai>=0.0.5
langchain>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.66.1